+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [feat]: `.distance()` and `.distanceSq()` methods;
 - [feat]: `angle()` method;
 - [feat]: `lerp()` method;
+- [feat]: `copy()` method;
 
 ## 0.6.0 (2022-11-27)